+++ gen/java/src/main/java/io/platform/contracts/events/ComponentHealthPayload.java
@@ -32,7 +32,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
   ComponentHealthPayload.JSON_PROPERTY_STATUS,
   ComponentHealthPayload.JSON_PROPERTY_DETAILS
 })
-@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2026-07-04T21:52:00.543013800+01:00[Africa/Casablanca]", comments = "Generator version: 7.23.0")
+@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2026-07-05T03:05:37.875350100+01:00[Africa/Casablanca]", comments = "Generator version: 7.23.0")
 public class ComponentHealthPayload {
   public static final String JSON_PROPERTY_COMPONENT_ID = "componentId";
   @jakarta.annotation.Nonnull